Move utilities out
